use sequelize init duplicate file delete
