+++ package.json
@@ -22,7 +22,9 @@
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
     "mysql": "^2.17.1",
+    "mysql2": "^1.7.0",
     "nodemon": "^1.19.4",
-    "sequelize": "^5.21.1"
+    "sequelize": "^5.21.1",
+    "sequelize-cli": "^5.5.1"
   }
 }
